+++ pyoti/multis/__init__.py
@@ -1,13 +1,22 @@
+from .binaryedge import BinaryEdge
 from .circlpssl import CIRCLPSSL
+from .ciscoumbrella import CiscoUmbrellaInvestigate
 from .dnsblocklist import DNSBlockList
+from .filescanio import FileScanIO
 from .hybridanalysis import HybridAnalysis
+from .ip2location import IP2Location, IP2WHOIS
+from .joesandbox import JoeSandbox
 from .maltiverseioc import MaltiverseIOC
+from .metadefendercloud import MetaDefenderCloudV4
 from .misp import MISP
 from .onyphe import Onyphe
 from .otx import OTX
 from .pulsedive import Pulsedive
+from .stairwell import Stairwell
 from .threatfox import ThreatFox
+from .triage import Triage
 from .urlhaus import URLhaus
 from .urlscan import URLscan
-from .virustotal import VirusTotalV2, VirusTotalV3
+from .virustotal import VirusTotalV3
+from .whoisxml import WhoisXML
 from .xforce import XForceExchange